+++ packages/ckb-sdk-rpc/types/CKB_RPC/index.d.ts
@@ -5,6 +5,7 @@
 /* eslint-disable camelcase */
 declare module CKB_RPC {
   export type ProposalShortId = CKBComponents.ProposalShortId
+  export type Number = CKBComponents.Number
   export type UInt32 = CKBComponents.UInt32
   export type Count = CKBComponents.Count
   export type Hash256 = CKBComponents.Hash256
@@ -50,6 +51,14 @@ declare module CKB_RPC {
   }
 
   export type Cell = CellOutput
+
+  export interface LiveCell {
+    data?: {
+      content: Hash
+      hash: Hash256
+    }
+    output: CellOutput
+  }
 
   export interface CellDep {
     out_point: OutPoint | null