+++ packages/ckb-sdk-utils/__tests__/serialization/script/index.test.js
@@ -8,38 +8,64 @@ const fixtures = require('./fixtures.json')
 
 describe('Test Script Serialization', () => {
   describe('Serialize Args', () => {
-    const fixtureTable = fixtures.serializeArgs.map(({ args, expected }) => [args, expected])
-    test.each(fixtureTable)('%j => %s', (args, expected) => {
-      expect(serializeArgs(args)).toBe(expected)
+    const fixtureTable = Object.entries(fixtures.serializeArgs).map(([title, { args, expected, exception }]) => [
+      title,
+      args,
+      expected,
+      exception,
+    ])
+    test.each(fixtureTable)('%s', (_title, args, expected, exception) => {
+      if (undefined !== expected) {
+        expect(serializeArgs(args)).toBe(expected)
+      }
+      if (undefined !== exception) {
+        expect(() => serializeArgs(args)).toThrow(new Error(exception))
+      }
     })
   })
 
   describe('Serialize CodeHash', () => {
-    const fixtureTable = fixtures.serializeCodeHash.map(({ codeHash, expected }) => [codeHash, expected])
-    test.each(fixtureTable)('%s => %s', (codeHash, expected) => {
-      expect(serializeCodeHash(codeHash)).toBe(expected)
+    const fixtureTable = Object.entries(fixtures.serializeCodeHash).map(
+      ([title, { codeHash, expected, exception }]) => [title, codeHash, expected, exception]
+    )
+    test.each(fixtureTable)('%s', (_title, codeHash, expected, exception) => {
+      if (undefined !== expected) {
+        expect(serializeCodeHash(codeHash)).toBe(expected)
+      }
+      if (undefined !== exception) {
+        expect(() => serializeCodeHash(codeHash)).toThrow(new Error(exception))
+      }
     })
   })
 
   describe('Serialize HashType', () => {
-    const fixtureTable = fixtures.serializeHashType.map(({ hashType, expected }) => [hashType, expected])
-    test.each(fixtureTable)('%s => %s', (hashType, expected) => {
-      expect(serializeHashType(hashType)).toBe(expected)
-    })
-
-    it('throw errors if hash type is neither data nor type', () => {
-      expect(() => serializeHashType('unknown')).toThrow(new TypeError("Hash type must be either of 'data' or 'type'"))
+    const fixtureTable = Object.entries(fixtures.serializeHashType).map(
+      ([title, { hashType, expected, exception }]) => [title, hashType, expected, exception]
+    )
+    test.each(fixtureTable)('%s', (_title, hashType, expected, exception) => {
+      if (undefined !== expected) {
+        expect(serializeHashType(hashType)).toBe(expected)
+      }
+      if (undefined !== exception) {
+        expect(() => serializeHashType(hashType)).toThrow(new Error(exception))
+      }
     })
   })
 
   describe('Serialize Script', () => {
-    const fixtureTable = fixtures.serializeScript.map(({ script, expected }) => [script, expected])
-    test.each(fixtureTable)('%j => %s', (script, expected) => {
-      expect(serializeScript(script)).toBe(expected)
-    })
-
-    it('throw errors if the script is not passed into the method', () => {
-      expect(() => serializeScript()).toThrow(new Error('Script is required'))
+    const fixtureTable = Object.entries(fixtures.serializeScript).map(([title, { script, expected, exception }]) => [
+      title,
+      script,
+      expected,
+      exception,
+    ])
+    test.each(fixtureTable)('%s', (_title, script, expected, exception) => {
+      if (undefined !== expected) {
+        expect(serializeScript(script)).toBe(expected)
+      }
+      if (undefined !== exception) {
+        expect(() => serializeScript(script)).toThrow(new Error(exception))
+      }
     })
   })
 })