+++ packages/ckb-sdk-utils/src/exceptions/hexStringShouldStartWith0x.ts
@@ -1,0 +1,5 @@
+export default class extends Error {
+  constructor(hex: string) {
+    super(`Hex string ${hex} should start with 0x`)
+  }
+}