+++ packages/ckb-sdk-utils/src/exceptions/index.ts
@@ -1,0 +1,11 @@
+import HexStringShouldStartWith0x from './hexStringShouldStartWith0x'
+import ArgumentRequired from './argumentRequired'
+import InvalidHexString from './invalidHexString'
+
+export { HexStringShouldStartWith0x, ArgumentRequired, InvalidHexString }
+
+export default {
+  HexStringShouldStartWith0x,
+  ArgumentRequired,
+  InvalidHexString,
+}